Give Filesystem what it needs to decide what's allowed
The current API doesn't have a way for a template to know whether to
call add_dependency() or read_text() for a built file that it wants to
read after the file is built. By moving the logic for what's allowed
into Filesystem, I can (in another commit) change read_text() to return
None for files that aren't readable yet but will be after they're built.