Make write_text() return whether it made changes or not
It looks like ninja only reloads subninja files when it reloads the main
build.ninja. Without restat on that, it reloads it in an infinite loop;
with restat, it only reloads when the mtime changes. This change lets me
track whether any subninja files changed, and update the main
build.ninja's mtime when any did.