Make escape_shell=False the default for to_ninja
When I wrote that function, I was thinking that there would be a lot
more calls with escape_shell=True than there ended up being. I also
thought that requiring an explicit value would help by making it
necessary to think about whether shell escaping is needed or not. In
practice though, I think it just became a mass of code that was easy to
gloss over and not notice whether it was True or False. Now calls with
escape_shell=True actually stand out, which I think will make it easier
to see that the code is correct.