Split Config into full and minimal sets of fields
This is very similar to filesystem._ConfigPaths and will mostly replace
that in another commit. I'm planning to write Minimal to a json file,
then use that file to trigger rebuilds of things that use it, rather
than not rebuilding when these fields change (current behavior) or
rebuilding when any field changes.