Make non-Minimal fields clearly preserve order
There isn't much benefit to hiding the order, since things that depend
on these will depend on the original config file.

While ginjarator itself shouldn't care about the order of these, I think
it is reasonable for templates to rely on the order. E.g., to filter
templates for ones matching "blog_post/*" and use the order in templates
as the order of blog posts.