Merge Mode.check_dependency and .check_output into read/write
The {add,check}_{dependency,output} concepts are currently somewhat
broken, since they can either mean "I'm reading/writing this now without
using the Filesystem API" or "I want to read/write this in another
pass". That's causing circular dependencies where scan-done.stamp
depends on a scan that depends on its deferred reads, and a deferred
read is also an output of another template's render pass, which depends
on scan-done.stamp. The solution is to make the scan pass only depend on
actual reads, not deferred reads.

This commit does not fix that problem, but I think it will make it
easier to in another commit. For now this is just a no-op refactor.