Remove restat, and test that updates trigger rebuild
With the old preserve_mtime/restat logic, these new e2e tests were
failing because the scan would happen in one ninja build, and the render
in the next build. I don't understand why, but removing the restat logic
seems to fix it and make the code simpler. Hopefully the performance
will be ok without restat.