Make scan-done depend on dyndep files, not ginjarator-internal files
Its purpose is to delay the render steps until after the dyndep files
from the scan steps are all read, so that the dynamic dependencies can
be built in the right order. In practice this change doesn't seem to
make a difference, but it seems more correct, I think. (I found this
while debugging some dynamic dependencies that weren't being updated
correctly. I'd hoped this would fix it, but it seems to have no effect.)